Move Factorio globals to their own file
In doing so they are more clearly marked as such and there is no confusion
on what their purpose is.
